modal component improve
